Change facts/fact list/facts to trivia/trivia list/quotes
