+++ _content/docs/trivia-list.md
@@ -1,0 +1,23 @@
+---
+layout: docs
+group: docs
+title: Trivia List
+sort: 98
+---
+
+A trivia list is a standard format for representing a list of trivia / anecdotes used in various places on the site.
+
+## <i class="fa fa-tags"></i> Attribute Reference
+
+{% include def-doc.html def=site.data.defs.trivia-list %}
+
+## <i class="octicon octicon-code"></i> Example Trivia List
+
+{% highlight yaml %}
+trivia:
+  - quote: Every character in this play was portrayed by a perfectly 
+          circular Victoria Sponge.
+    name: Fred Bloggs
+    submitted: 2016-01-01
+    
+{% endhighlight %}
