--- _includes/fact-list.html
@@ -1,45 +1,0 @@
-<ul class="fact-list">
-{% for fact in include.facts %}
-  <li class="fact-list_fact">
-    <i class="fa fa-comment-o fact-list_bullet"></i>
-    <span class="fact-list_quote">{{ fact.fact }}</span>
-    {% if fact.person %}
-      <span class="fact-list_dash">&mdash;</span>      
-      <span class="fact-list_person" 
-            {% if fact.submitted %}
-            title="Submitted: {{ fact.submitted }}"
-            {% endif %}>
-        <a href="{{ fact.person.url }}">
-          {% include images/person-headshot.html 
-                    image=fact.person.headshot
-                    sizes="1em"
-                    alt=fact.person.title %}
-          <span class="fact-list_name">{{ fact.person.title }}</span>
-        </a>
-      </span>
-    {% elsif fact.name %}
-      <span class="fact-list_dash">&mdash;</span>
-      <span class="fact-list_name" 
-            {% if fact.submitted %}
-            title="Submitted: {{ fact.submitted }}"
-            {% endif %}>
-        {{ fact.name }}
-      </span>
-
-    {% endif %}
-
-    {% if fact.submitted %}
-    <span class="fact-list_submitted"
-          title="{{ fact.submitted }}" 
-          data-date="{{ fact.submitted }}">
-      {{ fact.submitted }}
-    </span>
-    {% endif %}
-  </li>
-{% endfor %}
-
-  <li class="fact-list__add">
-    <i class="ion-plus fact-list_bullet"></i>
-    <a href="https://github.com/newtheatre/history-project/issues/new?body={{page.url}}" data-proofer-ignore data-report-this-page>Send in an anecdote</a>
-  </li>
-</ul>
