+++ _layouts/show.html
@@ -113,8 +113,8 @@ body_class: section-archives layout-show
       <article class="show-synopsys">
         {{ content }}
 
-        {% if show.facts %}
-          {% include fact-list.html facts=show.facts %}
+        {% if show.trivia %}
+          {% include trivia-list.html trivia=show.trivia %}
         {% elsif show.ignore_missing != true %}
           {% include boxes/show_facts_missing.html %}
         {% endif %}
