+++ _sass/_trivia-list.sass
@@ -1,4 +1,4 @@
-.fact-list
+.trivia-list
   // <ul>
 
   // Custom bullets
@@ -6,25 +6,25 @@
   list-style: none
   li
     position: relative
-  .fact-list_bullet
+  .trivia-list_bullet
     position: absolute
     left: -$baseline
     line-height: $baseline
 
-.fact-list_fact
+.trivia-list_item
   // List item 
 
-.fact-list_quote
+.trivia-list_quote
   // Actual quotation from submitter
   font-style: italic
   &:before, &:after
     content: '"'
 
-.fact-list_dash, .fact-list_name, .fact-list_submitted
+.trivia-list_dash, .trivia-list_name, .trivia-list_submitted
   // Make things grey
   color: $color-grey
 
-.fact-list_person
+.trivia-list_person
   // When matches person record, contains headshot and name
   .person-headshot-actual, .person-headshot-silhouette
     position: relative
@@ -42,18 +42,18 @@
   // Underline name only, not picture and space between
   a:hover
     text-decoration: none
-  &:hover .fact-list_name
+  &:hover .trivia-list_name
     text-decoration: underline
 
-.fact-list_name
+.trivia-list_name
   // When orphaned, and name within _person
 
-.fact-list_submitted
+.trivia-list_submitted
   &:before
     content: '('
   &:after
     content: ')'
 
-.fact-list__add
+.trivia-list__add
   a
     font-weight: normal
