+++ _shows/83_84/the_balcony.md
@@ -8,8 +8,8 @@ venue:
 date_start:
 date_end:
 
-facts:
-  - fact: "I had to hold an air pistol for the assassin role. The director wanted to make sure I'd feel comfortable handling it, so during a rehearsal in the Trent Building he insisted on me shooting a pellet into one of the walls of the tunnel to Portland Building. It was still there when I graduated a couple of years later."
+trivia:
+  - quote: "I had to hold an air pistol for the assassin role. The director wanted to make sure I'd feel comfortable handling it, so during a rehearsal in the Trent Building he insisted on me shooting a pellet into one of the walls of the tunnel to Portland Building. It was still there when I graduated a couple of years later."
     name: Vanessa Greatorex
     submitted: 2015-10-05
