+++ _shows/83_84/tom_thumb_the_great.md
@@ -8,8 +8,8 @@ venue:
 date_start:
 date_end:
 
-facts:
-  - fact: "This was hilarious casting, as Glumdalca is supposed to be a giantess and, at 5ft 3in, I was shorter than nearly all the characters I was supposed to dwarf. I had gastric flu the first two nights, but carried on regardless, complete with massively back-combed hair which took hours to untangle afterwards. My final line in the play, uttered after dramatically falling headlong, was 'I am dead.' I accidentally discovered that smoothing my skirt down decently after expiring would raise a laugh."
+trivia:
+  - quote: "This was hilarious casting, as Glumdalca is supposed to be a giantess and, at 5ft 3in, I was shorter than nearly all the characters I was supposed to dwarf. I had gastric flu the first two nights, but carried on regardless, complete with massively back-combed hair which took hours to untangle afterwards. My final line in the play, uttered after dramatically falling headlong, was 'I am dead.' I accidentally discovered that smoothing my skirt down decently after expiring would raise a laugh."
     name: Vanessa Greatorex
     submitted: 2015-10-05
