+++ _shows/99_00/the_rise_and_fall_of_little_voice.md
@@ -6,8 +6,8 @@ playwright: Jim Cartwright
 season_sort: 40
 venue: New Theatre
 
-facts:
-  - fact: "The show was set over 3 levels, obviously burnt down, and his love of pigeons was replaced by a love of lights, as a result there were many comments along the lines of 'we've never seen so many lights in the roof!' Much fun was had."
+trivia:
+  - quote: "The show was set over 3 levels, obviously burnt down, and his love of pigeons was replaced by a love of lights, as a result there were many comments along the lines of 'we've never seen so many lights in the roof!' Much fun was had."
     name: Adam Godden
     submitted: 2015-10-06
