+++ css/main.sass
@@ -57,9 +57,9 @@ $color-seasons: #134466
 $color-playwright: #9F3B15
 
 $color-debug: #4FFF1F
-$color-debug-attr-base: #219E11
+$color-debug-attr-base: #1e8c10
 $color-debug-attr-required: #E56818
-$color-debug-attr-generated: #0C82CC
+$color-debug-attr-generated: #0b5e92
 
 $grid-transition-enable: false
 $grid-column-count: 16
@@ -123,7 +123,7 @@ $octicons-font-path: "/lib/octicons/octicons/"
 @import "archives"
 @import "collect"
 @import "docs"
-@import "fact-list"
+@import "trivia-list"
 @import "fellows"
 @import "footer"
 @import "header"
